+++ src/commands/deploy.ts
@@ -6,11 +6,12 @@ import klaw = require('klaw');
 import { getAuraNameFromFileName } from '../parsers';
 import * as xml2js from 'xml2js';
 import * as fs from 'fs-extra';
-import { getAnyTTFromPath } from '../parsers/open';
 import { outputToString } from '../parsers/output';
 import { isEmptyUndOrNull, toArray } from '../util';
 import { DeployResult } from 'jsforce';
 import { FCCancellationToken, ForcecodeCommand } from './forcecodeCommand';
+import { IMetadataObject } from '../forceCode';
+import { getAnyTTMetadataFromPath } from '../parsers/getToolingType';
 
 export class DeployPackage extends ForcecodeCommand {
   constructor() {
@@ -178,22 +179,17 @@ export function createPackageXML(files: string[], lwcPackageXML?: string): Promi
       },
     };
     files.forEach(file => {
-      const fileTT: string | undefined = getAnyTTFromPath(file);
+      var fileTT: IMetadataObject | undefined = getAnyTTMetadataFromPath(file);
       if (!fileTT) {
         reject();
         return;
       }
       var member: string | undefined;
-      if (fileTT === 'AuraDefinitionBundle') {
+      if (fileTT.xmlName === 'AuraDefinitionBundle') {
         member = getAuraNameFromFileName(file, 'aura');
-      } else if (fileTT === 'LightningComponentBundle') {
+      } else if (fileTT.xmlName === 'LightningComponentBundle') {
         member = getAuraNameFromFileName(file, 'lwc');
-      } else if (
-        fileTT === 'Document' ||
-        fileTT === 'EmailTemplate' ||
-        fileTT === 'Report' ||
-        fileTT === 'Dashboard'
-      ) {
+      } else if (fileTT.inFolder) {
         const file2 = file.split(vscode.window.forceCode.projectRoot + path.sep).pop();
         member = file2 ? file2.substring(file2.indexOf(path.sep) + 1) : file2;
       } else {
@@ -203,14 +199,22 @@ export function createPackageXML(files: string[], lwcPackageXML?: string): Promi
       member = member && lIndexOfP > 0 ? member.substring(0, lIndexOfP) : member;
       if (member) {
         member = member.replace('\\', '/');
-        const index: number = findMDTIndex(packObj, fileTT);
+        const index: number = findMDTIndex(packObj, fileTT.xmlName);
+        const folderMeta = member.split('/')[0];
+        const memberIndex: number = index !== -1 ? findMemberIndex(packObj, index, folderMeta) : -1;
         if (index !== -1) {
           packObj.Package.types[index].members.push(member);
+          if (fileTT.inFolder && memberIndex === -1) {
+            packObj.Package.types[index].members.push(folderMeta);
+          }
         } else {
           var newMem: PXMLMember = {
             members: [member],
-            name: fileTT,
+            name: fileTT.xmlName,
           };
+          if (fileTT.inFolder) {
+            newMem.members.push(folderMeta);
+          }
           packObj.Package.types.push(newMem);
         }
       }
@@ -233,6 +237,12 @@ export function createPackageXML(files: string[], lwcPackageXML?: string): Promi
   function findMDTIndex(pxmlObj: PXML, type: string) {
     return pxmlObj.Package.types.findIndex(curType => {
       return curType.name === type;
+    });
+  }
+
+  function findMemberIndex(pxmlObj: PXML, index: number, member: string) {
+    return pxmlObj.Package.types[index].members.findIndex(curType => {
+      return curType === member;
     });
   }
 }