+++ src/commands/diff.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import * as parsers from './../parsers';
 import ForceCodeContentProvider from '../providers/ContentProvider';
 import { ForcecodeCommand } from './forcecodeCommand';
+import { getToolingTypeFromFolder } from '../parsers/getToolingType';
 
 const PROVIDER: string = 'forcecode://salesforce.com';
 
@@ -24,8 +25,8 @@ export class DiffMenu extends ForcecodeCommand {
     if (!vscode.window.activeTextEditor) {
       return;
     }
-    const ttype: string | undefined = parsers.getToolingType(
-      vscode.window.activeTextEditor.document
+    const ttype: string | undefined = getToolingTypeFromFolder(
+      vscode.window.activeTextEditor.document.uri
     );
     if (!ttype) {
       throw { message: 'Metadata type not supported for diffing' };
@@ -41,7 +42,7 @@ export default function diff(document: vscode.TextDocument, auraSource?: boolean
   if (!document) {
     return Promise.reject('No document open to diff with the server.');
   }
-  const toolingType: string | undefined = parsers.getToolingType(document);
+  const toolingType: string | undefined = getToolingTypeFromFolder(document.uri);
   const fileName: string | undefined = parsers.getWholeFileName(document);
   if (auraSource) {
     ForceCodeContentProvider.getInstance().auraSource = document;